fix: resolve all qsizetype/qint64 truncation warnings on iOS/macOS
Fix ~100 -Wshorten-64-to-32 warnings across 28 source files. Qt methods
like indexOf(), size(), count(), length() return qsizetype (long long on
64-bit) but were stored in int. Changed to qsizetype where the variable
stays in qsizetype context, or added static_cast<int>() where the value
must be passed to APIs expecting int.

+++ src/controllers/maincontroller.cpp
@@ -1984,7 +1984,7 @@ QVariantMap MainController::getFrameAt(int index) const {
 }
 
 int MainController::frameCount() const {
-    return m_currentProfile.steps().size();
+    return static_cast<int>(m_currentProfile.steps().size());
 }
 
 void MainController::createNewProfile(const QString& title) {
